[CLEANUP] #5723 Fix the URL encode issue for 6.2 and 7.6 in unit tests
When encoding an URL by the typolink() method, TYPO3 CMS 6.2
and TYPO3 CMS 7.6 have two different behaviors which depends
on the config setting 'uniqueLinkVars'.
When it is not set, a link will generated like foo[bar]=42
and foo%5Bbar%5D=42 otherwise..

Since TYPO3 7.6 the setting 'uniqueLinkVars' default behaviour and can
not be changed because the setting was removed.

To have the same result in both version, we now simulate the 7.6 behavior
in 6.2.

Change-Id: I522295dd543a7681bf6993c45d492039601766d6
Reviewed-on: https://review.typo3.org/48932
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>