[BUGFIX] Display registration statistics in the BE in all cases
Some organizers manually close registration for an event when a
certain number of registrations has been reached, and some organizers
close the waiting list when it's getting too full.

In those cases, it should still be possible to see both the regular
registrations as well as the waiting list registrations in the
BE module.

Fixes #2654